% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_reference.R
\name{extract_reference}
\alias{extract_reference}
\title{Extract reference labels}
\usage{
extract_reference(x)
}
\arguments{
\item{x}{A list containing the variables and their terms}
}
\value{
y A list containing the extracted reference term
}
\description{
Extracts the reference term for each variable in a logistic regression.
}
