% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggforest.R
\name{ggforest}
\alias{ggforest}
\title{Creates a forest plot}
\usage{
ggforest(m, intercept = FALSE, categories = NULL, ...)
}
\arguments{
\item{m}{A fitted logistic regression model object}

\item{intercept}{whether to include the intercept within the plot}

\item{...}{}
}
\value{
A forest plot
}
\description{
Creates a forest plot
}
\examples{
model <- glm(y ~ x, data=data, family=binomial)
ggforest(model)
}
