% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ref_df.R
\name{make_ref_df}
\alias{make_ref_df}
\title{Creates a tibble(data frame) with variables and their reference labels}
\usage{
make_ref_df(m)
}
\arguments{
\item{m}{A logistic regression model object}
}
\value{
A tibble(data frame) containing variable names and their reference labels
}
\description{
Creates a tibble(data frame) with variables and their reference labels
}
\examples{
\dontrun{
m <- glm(y ~ x, data=d, family=binomial)
make_ref_df(m)
}
}
