% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_samadult.R
\name{read_samadult_at}
\alias{read_samadult_at}
\title{Reads NHIS sample adult files for specified survey years}
\usage{
read_samadult_at(
  dir.path = NULL,
  years = NULL,
  lookup.tbl.path = NULL,
  rename = FALSE
)
}
\arguments{
\item{dir.path}{Path to the directory(folder) that contains NHIS sample adult files}

\item{years}{A vector of years}

\item{lookup.tbl.path}{Path to the variable name look-up file}

\item{rename}{Rename column names to user specified names}
}
\value{
A NHIS dataframe with only the variables specified in the variable name look-up file
}
\description{
Reads NHIS sample adult files for specified survey years
}
\examples{
\dontrun{
read_samadult_at(dir.path = "./samadult-files",
                 years = c(2006,2012,2015),
                 lookup.tbl.path = "./samadult-lookup.csv")
}
}
