% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggforest.R
\name{ggforest}
\alias{ggforest}
\title{Creates a forest plot}
\usage{
ggforest(
  model,
  include.intercept = FALSE,
  categories = NULL,
  color.scheme = NULL,
  title = NULL,
  xlab = "Terms",
  ylab = "Log OR",
  errorbar = TRUE,
  errorbar.width = 0.3,
  linetype = "dashed",
  point.size = 3,
  point.alpha = 1,
  ggtheme = theme_classic(),
  ...
)
}
\arguments{
\item{model}{A fitted logistic regression model object}

\item{include.intercept}{Whether to include the model intercept within the plot}

\item{categories}{A named list containing the grouping of terms (e.g. list(Race = c("RACE"), SocioDemographics = c("AGE", "SEX")))}

\item{color.scheme}{A named list containing the coloring for each category}

\item{title}{Plot title}

\item{xlab}{x label}

\item{ylab}{y label}

\item{errorbar}{Whether to draw errorbars}

\item{errorbar.width}{The width of the errorbars}

\item{linetype}{The linetype of the reference line}

\item{point.size}{Point size}

\item{point.alpha}{Alpha value of the points}

\item{ggtheme}{ggplot theme}

\item{...}{Other parameters passed to ggplot}
}
\value{
A forest plot
}
\description{
Creates a forest plot
}
\examples{
\dontrun{
model <- glm(y ~ x, data=data, family=binomial)
ggforest(model)
}
}
